// src/generated/model/verdierFrontend.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface VerdierFrontend {
    bekreftelse?: boolean;
    bolig: boolean;
    campingvogn: boolean;
    kjoretoy: boolean;
    fritidseiendom: boolean;
    annet: boolean;
    beskrivelseAvAnnet?: string;
}